rebase v1.0.0

+++ library-udf/src/main/java/org/apache/iotdb/library/anomaly/util/MissDetector.java
@@ -68,7 +68,7 @@ public class MissDetector {
     SimpleRegression regression = new SimpleRegression();
     while (i + windowSize < getLen()) {
       for (int j = 0; j < windowSize; j++) {
-        data[j][0] = (double) getTime().get(i + j) - startTime;
+        data[j][0] = getTime().get(i + j) - startTime;
         data[j][1] = getValue().get(i + j);
       }
       regression.addData(data);
@@ -91,19 +91,19 @@ public class MissDetector {
     int bindex = start;
     while (bindex > 0) {
       bindex--;
-      regression.addData((double) getTime().get(bindex) - startTime, getValue().get(bindex));
+      regression.addData(getTime().get(bindex) - startTime, getValue().get(bindex));
       double alpha = regression.getRSquare();
       if ((horizon && getValue().get(bindex) != standard) || (!horizon && alpha < threshold)) {
         break;
       }
     }
-    regression.removeData((double) getTime().get(bindex) - startTime, getValue().get(bindex));
+    regression.removeData(getTime().get(bindex) - startTime, getValue().get(bindex));
     if (bindex == 0) {
       return end;
     }
     int findex = end;
     while (findex < getLen()) {
-      regression.addData((double) getTime().get(findex) - startTime, getValue().get(findex));
+      regression.addData(getTime().get(findex) - startTime, getValue().get(findex));
       double alpha = regression.getRSquare();
       if ((horizon && getValue().get(findex) != standard) || (!horizon && alpha < threshold)) {
         break;
@@ -130,30 +130,22 @@ public class MissDetector {
     }
   }
 
-  /**
-   * @return the minLength
-   */
+  /** @return the minLength */
   public int getMinLength() {
     return minLength;
   }
 
-  /**
-   * @param minLength the minLength to set
-   */
+  /** @param minLength the minLength to set */
   public void setMinLength(int minLength) {
     this.minLength = minLength;
   }
 
-  /**
-   * @return the threshold
-   */
+  /** @return the threshold */
   public double getThreshold() {
     return threshold;
   }
 
-  /**
-   * @param threshold the threshold to set
-   */
+  /** @param threshold the threshold to set */
   public void setThreshold(double threshold) {
     this.threshold = threshold;
   }
@@ -184,44 +176,32 @@ public class MissDetector {
       return end - start;
     }
 
-    /**
-     * @return the start
-     */
+    /** @return the start */
     public int getStart() {
       return start;
     }
 
-    /**
-     * @param start the start to set
-     */
+    /** @param start the start to set */
     public void setStart(int start) {
       this.start = start;
     }
 
-    /**
-     * @return the end
-     */
+    /** @return the end */
     public int getEnd() {
       return end;
     }
 
-    /**
-     * @param end the end to set
-     */
+    /** @param end the end to set */
     public void setEnd(int end) {
       this.end = end;
     }
 
-    /**
-     * @return the slope
-     */
+    /** @return the slope */
     public double getSlope() {
       return slope;
     }
 
-    /**
-     * @param slope the slope to set
-     */
+    /** @param slope the slope to set */
     public void setSlope(double slope) {
       this.slope = slope;
     }
@@ -238,30 +218,22 @@ public class MissDetector {
     }
   }
 
-  /**
-   * @return the predictLabel
-   */
+  /** @return the predictLabel */
   public IntArrayList getPredictLabel() {
     return predictLabel;
   }
 
-  /**
-   * @return the len
-   */
+  /** @return the len */
   public int getLen() {
     return len;
   }
 
-  /**
-   * @return the time
-   */
+  /** @return the time */
   public LongArrayList getTime() {
     return time;
   }
 
-  /**
-   * @return the value
-   */
+  /** @return the value */
   public DoubleArrayList getValue() {
     return value;
   }
